requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
hl7==0.4.5
fhir.resources==7.1.0
python-multipart==0.0.6
requests==2.31.0
pydantic==2.5.0